Summary: Streaming voting machine
Voting machine that returns top three candidates and checks for voter Fraud on stream.

On election day, a voting machine writes data in the form (voter_id, candidate_id) to a text file. Write a program that reads this file as a stream and returns the top 3 candidates at any given time. If you find a voter voting more than once, report this as fraud.